Move file into right module
