+++ src/ui/setup_mainwindow.py
@@ -20,7 +20,7 @@ from src.updater import Updater
 from src.utils import has_connection
 
 from src.ui_elements import Ui_MainWindow
-from src.ui_elements.cocktail_view import CocktailView
+from src.ui.cocktail_view import CocktailView
 from src.ui.setup_option_window import OptionWindow
 from src.ui.setup_progress_screen import ProgressScreen
 from src.ui.setup_numpad_widget import NumpadWidget
